feat: fix user and change default copy behaviour
Set User at step level instead of stage level.
In this way we can easily swap between users when necessary.

Also changed the default llb.Copy behaviour, which copy the entire
directory to the destination, preserving the directory.
This is in contrast on what the standard Dockerfile do, which copy the
contents only